separa customização de classes em modulos
