+++ sapl/api/views_compilacao.py
@@ -1,0 +1,12 @@
+
+from django.apps.registry import apps
+
+from drfautoapi.drfautoapi import ApiViewSetConstrutor, \
+    customize, wrapper_queryset_response_for_drf_action
+
+
+CompilacaoApiViewSetConstrutor = ApiViewSetConstrutor.build_class(
+    [
+        apps.get_app_config('compilacao')
+    ]
+)
